Quarantine the pre-redesign research UI in ui/legacy
Existing conversations must keep rendering exactly as they were written,
so the old cards are moved out of ui/components into a frozen ui/legacy
package rather than edited in place: LegacyResearchProgressCard,
LegacyReportCard, LegacyResearchPlanDisclosure, LegacyDataResultCard, and
a frozen copy of SourcesRow.

Nothing in ui/components depends on the legacy package, and it is reached
from a single dispatch point in ChatMessages. The file carries a note
saying not to improve anything in it -- a change to these composables
retroactively rewrites how old chats look, which is the exact failure the
split exists to prevent.

The one thing legacy still inherits is the design system itself: theme
colours, typography, Spacing, markdown rendering. Old messages should keep
following palette and font changes or they would look broken sitting next
to new ones. Composition is frozen; skin is not.

ResearchComponents.kt keeps only the composer-side controls (EffortPill,
CapabilityChip), which were never part of rendering a run.
