+++ app/src/main/java/com/echoflow/ui/screens/ChatComposer.kt
@@ -107,12 +107,9 @@ import com.echoflow.ui.components.ArtifactCard
 import com.echoflow.ui.components.CapabilityChip
 import com.echoflow.ui.components.ContextChipRow
 import com.echoflow.ui.components.ModelPill
-import com.echoflow.ui.components.DataResultCard
 import com.echoflow.ui.components.FusionCard
 import com.echoflow.ui.components.EffortPill
 import com.echoflow.ui.components.MarkdownText
-import com.echoflow.ui.components.ReportCard
-import com.echoflow.ui.components.ResearchProgressCard
 import com.echoflow.ui.components.RichMarkdown
 import com.echoflow.ui.components.SearchActivityCard
 import com.echoflow.ui.components.SectionLabel
